Split branch creation and deletion into their own home
