+++ crates/gitbutler-branch/src/branch_ext.rs
@@ -1,5 +1,5 @@
 use anyhow::{Context, Result};
-use gitbutler_reference::ReferenceName;
+use gitbutler_tagged_string::ReferenceName;
 
 pub trait BranchExt {
     fn reference_name(&self) -> Result<ReferenceName>;
