+++ crates/gitbutler-branch/src/lib.rs
@@ -1,5 +1,6 @@
 pub mod branch;
 pub mod branch_ext;
+pub mod dedup;
 pub mod diff;
 pub mod file_ownership;
 pub mod hunk;
