+++ crates/gitbutler-reference/src/lib.rs
@@ -1,5 +1,4 @@
 mod refname;
-use gitbutler_tagged_string::TaggedString;
 pub use refname::{LocalRefname, Refname, RemoteRefname, VirtualRefname};
 use regex::Regex;
 
@@ -7,7 +6,3 @@ pub fn normalize_branch_name(name: &str) -> String {
     let pattern = Regex::new("[^A-Za-z0-9_/.#]+").unwrap();
     pattern.replace_all(name, "-").to_string()
 }
-
-pub struct _ReferenceName;
-/// The name of a reference ie. `refs/heads/master`
-pub type ReferenceName = TaggedString<_ReferenceName>;
