+++ crates/gitbutler-tauri/Cargo.toml
@@ -32,7 +32,7 @@ git2.workspace = true
 once_cell = "1.19"
 reqwest = { version = "0.12.4", features = ["json"] }
 serde.workspace = true
-serde_json = { version = "1.0", features = [ "std", "arbitrary_precision" ] }
+serde_json = { version = "1.0", features = ["std", "arbitrary_precision"] }
 tauri-plugin-context-menu = { git = "https://github.com/c2r0b/tauri-plugin-context-menu", branch = "main" }
 tauri-plugin-single-instance = { git = "https://github.com/tauri-apps/plugins-workspace", branch = "v1" }
 tauri-plugin-window-state = { git = "https://github.com/tauri-apps/plugins-workspace", branch = "v1" }
@@ -41,7 +41,7 @@ tauri-plugin-log = { git = "https://github.com/tauri-apps/plugins-workspace", br
 log = "^0.4"
 thiserror.workspace = true
 # The features here optimize for performance.
-tokio = { workspace = true, features = [ "rt-multi-thread", "parking_lot" ] }
+tokio = { workspace = true, features = ["rt-multi-thread", "parking_lot"] }
 tracing = "0.1.40"
 tracing-appender = "0.2.3"
 tracing-subscriber = "0.3.17"
@@ -60,15 +60,23 @@ gitbutler-error.workspace = true
 gitbutler-secret.workspace = true
 gitbutler-id.workspace = true
 gitbutler-storage.workspace = true
+gitbutler-tagged-string.workspace = true
 open = "5"
 
 [dependencies.tauri]
 version = "1.7.0"
 features = [
-    "http-all", "os-all", "dialog-open", "fs-read-file",
-    "path-all", "process-relaunch", "protocol-asset",
-    "shell-open", "window-maximize", "window-start-dragging",
-    "window-unmaximize"
+    "http-all",
+    "os-all",
+    "dialog-open",
+    "fs-read-file",
+    "path-all",
+    "process-relaunch",
+    "protocol-asset",
+    "shell-open",
+    "window-maximize",
+    "window-start-dragging",
+    "window-unmaximize",
 ]
 
 [lints.clippy]
@@ -86,4 +94,4 @@ devtools = ["tauri/devtools"]
 # DO NOT remove this
 custom-protocol = ["tauri/custom-protocol"]
 
-error-context = ["dep:backtrace" ]
+error-context = ["dep:backtrace"]
