+++ crates/gitbutler-tauri/src/virtual_branches.rs
@@ -6,8 +6,8 @@ pub mod commands {
     use gitbutler_error::error::Code;
     use gitbutler_project as projects;
     use gitbutler_project::ProjectId;
-    use gitbutler_reference::ReferenceName;
     use gitbutler_reference::{Refname, RemoteRefname};
+    use gitbutler_tagged_string::ReferenceName;
     use gitbutler_virtual::assets;
     use gitbutler_virtual::base::BaseBranch;
     use gitbutler_virtual::files::RemoteBranchFile;
