+++ crates/gitbutler-virtual/Cargo.toml
@@ -22,7 +22,8 @@ gitbutler-id.workspace = true
 gitbutler-time.workspace = true
 gitbutler-commit.workspace = true
 gitbutler-url.workspace = true
-serde = { workspace = true, features = ["std"]}
+gitbutler-tagged-string.workspace = true
+serde = { workspace = true, features = ["std"] }
 bstr = "1.9.1"
 diffy = "0.3.0"
 hex = "0.4.3"
@@ -38,7 +39,7 @@ urlencoding = "2.1.3"
 reqwest = { version = "0.12.4", features = ["json"] }
 
 [[test]]
-name="virtual"
+name = "virtual"
 path = "tests/virtual_branches/mod.rs"
 
 [dev-dependencies]
